Added profilers and electricity meters
profiler.py contains pytorch profiler and cProfiler decorator that can be used with the function that has to be tested (perform_training, perform_validation, fit).
Also, energy_profiling decorator function measure the amount of energy used by hardware and amount of CO(2) that was emitted during the decorated function execution.